requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-huggingface==0.1.2
langchain-community==0.3.7
faiss-cpu==1.9.0
sentence-transformers==3.3.1
